workspace/py-workspace/testing-focus-session/requirements.txt
jupyter
pytest
pytest-parametrization
mock
